refactor(timeline): extract quick selector panel into separate component
- Created new QuickSelectorPanel component
- Moved panel rendering logic out of TimelineSelector
- Extracted triangle menu interaction logic into a separate hook
- Simplified TimelineSelector component structure
- Exported useTriangleMenu hook from hooks package
