+++ packages/hooks/src/index.ts
@@ -15,5 +15,6 @@ export * from "./useRefValue"
 export * from "./useSetState"
 export * from "./useSyncTheme"
 export * from "./useTitle"
+export * from "./useTriangleMenu"
 export * from "./useTypescriptHappyCallback"
 export * from "./useVideo"
